requirements.txt
albumentations==0.4.5
hydra-core==1.0.3
opencv-python==4.2.0.34
pytorch-lightning==1.2.2
torch==1.7.0
torchvision
tqdm==4.46.1
matplotlib
split-folders
scikit-learn